+++ src/heal_swin_nnx/models/swin.py
@@ -46,9 +46,9 @@ class SwinParams:
 
     # precision
     param_dtype: str = "float32"     # parameter storage; any DTypeLike, stored as name
-    dtype: str = "float32"           # compute/matmul dtype; "float32" is a staging
-                                     # default — flipped to "bfloat16" in the final
-                                     # task of the compute-dtype plan
+    dtype: str = "bfloat16"          # compute/matmul dtype; fp32 islands (norms,
+                                     # softmax, RoPE, final projections) are
+                                     # knob-independent — see the compute-dtype spec
 
     def __post_init__(self):
         self.img_size = _pair(self.img_size)
@@ -284,7 +284,7 @@ class SwinBlock(nnx.Module):
 
 
 class PatchMerging(nnx.Module):
-    def __init__(self, input_resolution, dim, param_dtype="float32", dtype="float32",
+    def __init__(self, input_resolution, dim, param_dtype="float32", dtype="bfloat16",
                  *, rngs):
         self.input_resolution = tuple(input_resolution)
         self.reduction = nnx.Linear(4 * dim, 2 * dim, use_bias=False,
@@ -309,7 +309,7 @@ class PatchMerging(nnx.Module):
 
 class PatchExpand(nnx.Module):
     def __init__(self, input_resolution, dim, dim_scale=2, param_dtype="float32",
-                 dtype="float32", *, rngs):
+                 dtype="bfloat16", *, rngs):
         self.input_resolution = tuple(input_resolution)
         self.dtype = dtype
         self.expand = (nnx.Linear(dim, 2 * dim, use_bias=False, kernel_init=TRUNC_NORMAL,
@@ -332,7 +332,7 @@ class PatchExpand(nnx.Module):
 
 class FinalPatchExpand(nnx.Module):
     def __init__(self, input_resolution, patch_size, dim, param_dtype="float32",
-                 dtype="float32", *, rngs):
+                 dtype="bfloat16", *, rngs):
         self.input_resolution = tuple(input_resolution)
         self.patch_size = tuple(patch_size)
         self.output_dim = dim
